Adjust argument info to limit BC breaks.
Most of those things had to be adjusted because the documentation
was imcomplete or incorrect, which I've fixed as well in the PHP Documentation
SVN repository. There are still a few BC breaks; in the case of the options
array for insert/remote/update. Only an array is allowed now and not a single
boolean value. It wasn't really documented, and already deprecated in 1.0.5.

I've also changed/broken up some PHPUnit test cases.