Removed the "fireAndForget" connection string option,
I have also renamed default_do_fire_and_forget to default_do_gle as that makes
more sense.  This is probably not done ,but good enough until the mess has
sorted itself out for 1.4. Sadly, it means people can not turn on "safe" by
default yet on the connection object in this release.