Fixed the connection fetching for replicasets.
To make sure we only retrieve connections from the manager from the same
replicaset, we now store the replicaset set name as part of the server
definition as well as the hash. Then after fetching all candidate connections,
we filter on the replicaset name as stored in the hash and compare it to what
the parsed information contains.